requirements.txt
asgiref==3.2.3
Django==3.0
djangorestframework==3.11.0
gunicorn==20.0.4
phonenumbers==8.11.1
pytz==2019.1
sqlparse==0.3.0